Add sensor for Transmission port status
Just committing as-is for now with a basic sensor.
I'm not sure what is involved in checking the port status on the
part of the transmission daemon itself, but it seems like it's a
fairly intensive operation, more than the other sensors that are
mostly just scraping data. Could look into making this configurable
to either not run at all unless monitored, or run at a different
interval. Both would be best for my use case so I can turn it on
and have it only run once per hour.